Splitted NavigatorComposition.
