+++ ts/Core/Navigator.ts
@@ -42,6 +42,7 @@ const {
 import D from './DefaultOptions.js';
 const { defaultOptions } = D;
 import NavigatorAxisAdditions from './Axis/NavigatorAxisComposition.js';
+import NavigatorComposition from './Navigator/NavigatorComposition.js';
 import { Palette } from './Color/Palettes.js';
 import RendererRegistry from './Renderer/RendererRegistry.js';
 import Scrollbar from './Scrollbar.js';
@@ -845,6 +846,19 @@ RendererRegistry.getRendererType().prototype.symbols[
  *        Chart object
  */
 class Navigator {
+    public static compose(
+        AxisClass: typeof Axis,
+        ChartClass: typeof Chart,
+        SeriesClass: typeof Series
+    ): void {
+        NavigatorComposition.compose(
+            AxisClass,
+            ChartClass,
+            Navigator,
+            SeriesClass
+        );
+    }
+
     public constructor(chart: Chart) {
         this.init(chart);
     }
@@ -2771,175 +2785,7 @@ class Navigator {
 }
 // End of prototype
 
-if (!H.Navigator) {
-    H.Navigator = Navigator as any;
-
-    NavigatorAxisAdditions.compose(Axis);
-
-    // For Stock charts. For x only zooming, do not to create the zoom button
-    // because X axis zooming is already allowed by the Navigator and Range
-    // selector. (#9285)
-    addEvent(Chart, 'beforeShowResetZoom', function (): (boolean|undefined) {
-        const chartOptions = this.options,
-            navigator = chartOptions.navigator,
-            rangeSelector = chartOptions.rangeSelector;
-
-        if (((navigator && navigator.enabled) ||
-            (rangeSelector && rangeSelector.enabled)) &&
-            ((!isTouchDevice &&
-            (chartOptions.chart as any).zooming.type === 'x') ||
-            (isTouchDevice &&
-            (chartOptions.chart as any).zooming.pinchType === 'x'))
-        ) {
-            return false;
-        }
-    });
-
-    // Initialize navigator for stock charts
-    addEvent(Chart, 'beforeRender', function (): void {
-        const options = this.options;
-
-        if ((options.navigator as any).enabled ||
-            (options.scrollbar as any).enabled
-        ) {
-            this.scroller = this.navigator = new Navigator(this);
-        }
-    });
-
-    // For stock charts, extend the Chart.setChartSize method so that we can set
-    // the final top position of the navigator once the height of the chart,
-    // including the legend, is determined. #367. We can't use Chart.getMargins,
-    // because labels offsets are not calculated yet.
-    addEvent(Chart, 'afterSetChartSize', function (): void {
-
-        let legend = this.legend,
-            navigator = this.navigator,
-            scrollbarHeight,
-            legendOptions,
-            xAxis,
-            yAxis;
-
-        if (navigator) {
-            legendOptions = legend && legend.options;
-            xAxis = navigator.xAxis;
-            yAxis = navigator.yAxis;
-            scrollbarHeight = navigator.scrollbarHeight;
-
-            // Compute the top position
-            if (this.inverted) {
-                navigator.left = navigator.opposite ?
-                    this.chartWidth - (scrollbarHeight as any) -
-                    navigator.height :
-                    this.spacing[3] + (scrollbarHeight as any);
-                navigator.top = this.plotTop + (scrollbarHeight as any);
-            } else {
-                navigator.left = pick(
-                    xAxis.left,
-                    this.plotLeft + (scrollbarHeight as any)
-                );
-                navigator.top = (navigator.navigatorOptions.top as any) ||
-                    this.chartHeight -
-                    navigator.height -
-                    (scrollbarHeight as any) -
-                    this.spacing[2] -
-                    (
-                        this.rangeSelector && this.extraBottomMargin ?
-                            this.rangeSelector.getHeight() :
-                            0
-                    ) -
-                    (
-                        (
-                            legendOptions &&
-                            legendOptions.verticalAlign === 'bottom' &&
-                            legendOptions.layout !== 'proximate' && // #13392
-                            legendOptions.enabled &&
-                            !legendOptions.floating
-                        ) ?
-                            legend.legendHeight +
-                            pick(legendOptions.margin, 10) :
-                            0
-                    ) -
-                    (
-                        this.titleOffset ? this.titleOffset[2] : 0
-                    );
-            }
-
-            if (xAxis && yAxis) { // false if navigator is disabled (#904)
-
-                if (this.inverted) {
-                    xAxis.options.left = yAxis.options.left = navigator.left;
-                } else {
-                    xAxis.options.top = yAxis.options.top = navigator.top;
-                }
-
-                xAxis.setAxisSize();
-                yAxis.setAxisSize();
-            }
-        }
-    });
-
-    // Merge options, if no scrolling exists yet
-    addEvent(Chart, 'update', function (e: Chart): void {
-
-        const navigatorOptions = (e.options.navigator || {}),
-            scrollbarOptions = (e.options.scrollbar || {});
-
-        if (!this.navigator && !this.scroller &&
-            (navigatorOptions.enabled || scrollbarOptions.enabled)
-        ) {
-            merge(true, this.options.navigator, navigatorOptions);
-            merge(true, this.options.scrollbar, scrollbarOptions);
-            delete e.options.navigator;
-            delete e.options.scrollbar;
-        }
-
-    });
-
-    // Initialize navigator, if no scrolling exists yet
-    addEvent(Chart, 'afterUpdate', function (
-        event: Chart.AfterUpdateEventObject
-    ): void {
-
-        if (!this.navigator && !this.scroller &&
-            ((this.options.navigator as any).enabled ||
-            (this.options.scrollbar as any).enabled)
-        ) {
-            this.scroller = this.navigator = new (Navigator as any)(this);
-
-            if (pick(event.redraw, true)) {
-                this.redraw(event.animation); // #7067
-            }
-        }
-
-    });
-
-    // Handle adding new series
-    addEvent(Chart, 'afterAddSeries', function (): void {
-        if (this.navigator) {
-            // Recompute which series should be shown in navigator, and add them
-            this.navigator.setBaseSeries(null as any, false);
-        }
-    });
-
-    // Handle updating series
-    addEvent(Series, 'afterUpdate', function (): void {
-        if (this.chart.navigator && !this.options.isInternal) {
-            this.chart.navigator.setBaseSeries(null as any, false);
-        }
-    });
-
-    Chart.prototype.callbacks.push(function (chart: Chart): void {
-        let extremes,
-            navigator = chart.navigator;
-
-        // Initialize the navigator
-        if (navigator && chart.xAxis[0]) {
-            extremes = chart.xAxis[0].getExtremes();
-            navigator.render(extremes.min, extremes.max);
-        }
-    });
-}
-
+Navigator.compose(Axis, Chart, Series);
 H.Navigator = Navigator;
 
 declare module './Renderer/SVG/SymbolType' {
@@ -2948,4 +2794,10 @@ declare module './Renderer/SVG/SymbolType' {
     }
 }
 
-export default H.Navigator;
+/* *
+ *
+ *  Default Export
+ *
+ * */
+
+export default Navigator;
